requirements.txt
Flask==2.3.3
tensorflow
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
Werkzeug==2.3.7
gunicorn==21.2.0
